requirements.txt
tensorflow-gpu==1.13.1
scipy==1.4.1
numpy==1.16.2
